pyproject.toml
[tool.poetry]
name = "advent-of-code"
version = "0.1.0"
description = ""
authors = ["sheldon"]
license = "mit"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.1.3"
pandas = "^2.2.3"
advent-of-code-data = "^2.0.4"
pytest = "^8.3.4"
parse = "^1.20.2"
advent-of-code-ocr = "^1.0.0"
colorama = "^0.4.6"
rich = "^13.9.4"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
google-generativeai = "^0.8.3"
openai = "^1.56.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
